requirements.txt
# CORE LANGCHAIN & AGENTS
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-ollama>=0.1.0
langchain-groq>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.0.20

# DATABASE & ORM
sqlalchemy>=2.0.0
alembic>=1.13.0

# VECTOR STORES & EMBEDDINGS
chromadb>=0.4.0
langchain-chroma>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# ACADEMIC APIs & PAPER SOURCES
arxiv>=2.1.0
semanticscholar>=0.6.0
scholarly>=1.7.0

# PDF PROCESSING
pypdf>=3.17.0
pdfplumber>=0.10.0
pymupdf>=1.23.0

# DATA PROCESSING & ANALYSIS
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# WEB SCRAPING
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0
lxml>=5.0.0

# NLP & TEXT PROCESSING
spacy>=3.7.0
nltk>=3.8.0

# DOCUMENT CONVERSION
markdown>=3.5.0

# VISUALIZATION
matplotlib>=3.8.0
seaborn>=0.12.0
networkx>=3.2.0
plotly>=5.18.0

# WEB INTERFACE (OPTIONAL)
streamlit>=1.29.0
streamlit-aggrid>=0.3.4

# CONFIGURATION & ENVIRONMENT
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM API CLIENTS
openai>=1.6.0
anthropic>=0.8.0

# UTILITIES
tqdm>=4.66.0
rich>=13.7.0
click>=8.1.0

# TESTING (OPTIONAL)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# CODE QUALITY (OPTIONAL)
black>=23.12.0
flake8>=6.1.0
mypy>=1.8.0

# DOCUMENTATION (OPTIONAL)
mkdocs>=1.5.0
mkdocs-material>=9.5.0

# FASTAPI BACKEND (REQUIRED)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6
httpx>=0.26.0

